fix linter issues
